- Moved common classed to own library - Removed unused IRepository interface
