+++ Core/Solution.cs
@@ -1,4 +1,4 @@
-﻿namespace SkyrimCompileHelper.Common
+﻿namespace SkyrimCompileHelper.Core
 {
     using System.Collections.Generic;
