+++ Core/packages.config
@@ -1,0 +1,7 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Fody" version="1.29.0" targetFramework="net45" developmentDependency="true" />
+  <package id="Newtonsoft.Json" version="7.0.1" targetFramework="net45" />
+  <package id="PropertyChanged.Fody" version="1.50.1" targetFramework="net45" developmentDependency="true" />
+  <package id="semver" version="1.1.2" targetFramework="net45" />
+</packages>
