+++ SkyrimCompileHelper.sln
@@ -7,6 +7,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SkyrimCompileHelper", "Skyr
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Compiler", "Compiler\Compiler.csproj", "{AA767A0F-F961-4F68-A38E-F9BEED8BAD77}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Core", "Core\Core.csproj", "{4D07678A-BFD9-4D7B-8625-9AAC541BB6F1}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{AA767A0F-F961-4F68-A38E-F9BEED8BAD77}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AA767A0F-F961-4F68-A38E-F9BEED8BAD77}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AA767A0F-F961-4F68-A38E-F9BEED8BAD77}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4D07678A-BFD9-4D7B-8625-9AAC541BB6F1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4D07678A-BFD9-4D7B-8625-9AAC541BB6F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4D07678A-BFD9-4D7B-8625-9AAC541BB6F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4D07678A-BFD9-4D7B-8625-9AAC541BB6F1}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
