--- SkyrimCompileHelper/Common/IRepository.cs
@@ -1,15 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SkyrimCompileHelper.Common
-{
-    public interface IRepository<T>
-    {
-        IEnumerable<T> GetAll();
-
-        T GetByIdentfier();
-    }
-}
