+++ SkyrimCompileHelper/ViewModels/ChangeVersionViewModel.cs
@@ -21,9 +21,7 @@ namespace SkyrimCompileHelper.ViewModels
     [ImplementPropertyChanged]
     public class ChangeVersionViewModel : Screen
     {
-        private readonly SemVersion oldVersion;
-
-        /// <summary>Initializes a new instance of the <see cref="ChangeVersionViewModel"/> class.</summary>
+        /// <summary>Initialises a new instance of the <see cref="ChangeVersionViewModel"/> class.</summary>
         public ChangeVersionViewModel()
         {
             if (Execute.InDesignMode)
@@ -38,8 +36,6 @@ namespace SkyrimCompileHelper.ViewModels
 
         public ChangeVersionViewModel(SemVersion version)
         {
-            this.oldVersion = version;
-
             this.Major = version.Major;
             this.Minor = version.Minor;
             this.Patch = version.Patch;
