+++ SkyrimCompileHelper/ViewModels/ShellViewModel.cs
@@ -23,7 +23,8 @@ namespace SkyrimCompileHelper.ViewModels
 
     using PropertyChanged;
 
-    using SkyrimCompileHelper.Common;
+    using SkyrimCompileHelper.Core;
+    using SkyrimCompileHelper.Core.Repositories;
 
     /// <summary>Provides methods and properties to show the shell view.</summary>
     [ImplementPropertyChanged]
