+++ SkyrimCompileHelper/ViewModels/SolutionManagerViewModel.cs
@@ -17,7 +17,7 @@ namespace SkyrimCompileHelper.ViewModels
 
     using Caliburn.Micro;
 
-    using SkyrimCompileHelper.Common;
+    using SkyrimCompileHelper.Core;
 
     /// <summary>ViewModel containing methods and properties to add new solutions to the program.</summary>
     public sealed class SolutionManagerViewModel : Screen
