+++ SkyrimCompileHelper/ViewModels/SolutionViewModel.cs
@@ -24,8 +24,9 @@ namespace SkyrimCompileHelper.ViewModels
 
     using Semver;
 
-    using SkyrimCompileHelper.Common;
+    using SkyrimCompileHelper.Core;
     using SkyrimCompileHelper.Compiler;
+    using SkyrimCompileHelper.Core.Repositories;
 
     /// <summary>ViewModel containing methods and properties to work with a solution.</summary>
     [ImplementPropertyChanged]
