+++ SkyrimCompileHelper/Views/NewSolutionView.xaml
@@ -6,12 +6,12 @@
         xmlns:vm="clr-namespace:SkyrimCompileHelper.ViewModels"
         xmlns:cal="http://www.caliburnproject.org"
         xmlns:i="http://schemas.microsoft.com/expression/2010/interactivity"
-        xmlns:converter="clr-namespace:SkyrimCompileHelper.Common.Converter"
+        xmlns:converter1="clr-namespace:SkyrimCompileHelper.Core.Converter;assembly=Core"
         mc:Ignorable="d"
         d:DataContext="{d:DesignInstance Type=vm:NewSolutionViewModel, IsDesignTimeCreatable=True}"
         cal:Bind.AtDesignTime="True">
     <UserControl.Resources>
-        <converter:SemVersionToStringConverter x:Key="SemVersionToStringConverter" />
+        <converter1:SemVersionToStringConverter x:Key="SemVersionToStringConverter" />
     </UserControl.Resources>
 
     <Grid Margin="5">
